backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
httpx==0.27.0
python-dotenv==1.0.1
PyJWT==2.8.0
